refactor: add get_entry_meta for inspector proxy
