+++ src/promptflow-devkit/promptflow/_proxy/_errors.py
@@ -1,0 +1,8 @@
+# ---------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# ---------------------------------------------------------
+from promptflow.exceptions import SystemErrorException
+
+
+class ExecutorServiceUnhealthy(SystemErrorException):
+    pass
