+++ src/promptflow-devkit/promptflow/batch/_errors.py
@@ -2,7 +2,7 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # ---------------------------------------------------------
 
-from promptflow.exceptions import ErrorTarget, SystemErrorException, UserErrorException, ValidationException
+from promptflow.exceptions import ErrorTarget, UserErrorException, ValidationException
 
 
 class InputMappingError(ValidationException):
@@ -11,10 +11,6 @@ class InputMappingError(ValidationException):
 
 
 class EmptyInputsData(UserErrorException):
-    pass
-
-
-class ExecutorServiceUnhealthy(SystemErrorException):
     pass
