+++ src/promptflow/tests/test_configs/flows/saved/.gitignore
@@ -1,2 +1,1 @@
-# TODO: hide all the files in the directory
-# *
+*
