--- src/promptflow/tests/test_configs/flows/saved/hello_world_main/flow.dag.yaml
@@ -1,12 +1,0 @@
-entry: hello:hello_world
-environment:
-  python_requirements_txt: requirements
-  image: python:3.8-slim
-inputs:
-  text:
-    type: str
-    description: The text to be printed
-outputs:
-  answer:
-    type: str
-    description: The answer
