--- src/promptflow/tests/test_configs/flows/saved/hello_world_main/hello.py
@@ -1,4 +1,0 @@
-
-
-def hello_world(text: str) -> str:
-    return f"Hello World {text}!"
