--- src/promptflow/tests/test_configs/flows/saved/hello_world_main/requirements
@@ -1,1 +1,0 @@
-promptflow
