--- src/promptflow/tests/test_configs/flows/saved/hello_world_main/sample.json
@@ -1,3 +1,0 @@
-{
-    "text": "promptflow"
-}
